move the base functionality to separate preprocessing modules
