+++ h2integrate/feedstocks/eia_ng_pricing.py
@@ -1,4 +1,3 @@
-import os
 import json
 from pathlib import Path
 from datetime import datetime
@@ -17,182 +16,7 @@ from h2integrate.core.model_baseclasses import BaseConfig, CostModelBaseClass
 
 HOURS_PER_YEAR = 8760
 SECONDS_PER_HOUR = 3600
-MCF_to_MMBTU = 1 / 0.964
-STATE_MAP = {
-    "Alabama": "AL",
-    "Alaska": "AK",
-    "Arizona": "AZ",
-    "Arkansas": "AR",
-    "California": "CA",
-    "Colorado": "CO",
-    "Connecticut": "CT",
-    "Delaware": "DE",
-    "Florida": "FL",
-    "Georgia": "GA",
-    "Hawaii": "HI",
-    "Idaho": "ID",
-    "Illinois": "IL",
-    "Indiana": "IN",
-    "Iowa": "IA",
-    "Kansas": "KS",
-    "Kentucky": "KY",
-    "Louisiana": "LA",
-    "Maine": "ME",
-    "Maryland": "MD",
-    "Massachusetts": "MA",
-    "Michigan": "MI",
-    "Minnesota": "MN",
-    "Mississippi": "MS",
-    "Missouri": "MO",
-    "Montana": "MT",
-    "Nebraska": "NE",
-    "Nevada": "NV",
-    "New Hampshire": "NH",
-    "New Jersey": "NJ",
-    "New Mexico": "NM",
-    "New York": "NY",
-    "North Carolina": "NC",
-    "North Dakota": "ND",
-    "Ohio": "OH",
-    "Oklahoma": "OK",
-    "Oregon": "OR",
-    "Pennsylvania": "PA",
-    "Rhode Island": "RI",
-    "South Carolina": "SC",
-    "South Dakota": "SD",
-    "Tennessee": "TN",
-    "Texas": "TX",
-    "Utah": "UT",
-    "Vermont": "VT",
-    "Virginia": "VA",
-    "Washington": "WA",
-    "West Virginia": "WV",
-    "Wisconsin": "WI",
-    "Wyoming": "WY",
-    "District of Columbia": "DC",
-    "United States": "US",
-}
 CURRENT_YEAR = datetime.now().year
-
-EIA_FACET = {
-    "wellhead": "N9190{}3",
-    "imports": "N9100{}3",
-    "citygate": "N3050{}3",
-    "residential": "N3010{}3",
-    "commercial": "N3020{}3",
-    "industrial": "N3035{}3",
-    "electrical_power": "N3045{}3",
-    "exports": "N9130{}3",
-}
-
-
-def convert_to_monthly(df: pd.DataFrame) -> pd.DataFrame | None:
-    """Converts an annual timeseries to monthly by repeating the one value, or returns
-    the data passed, if already monthly.
-
-    Args:
-        df (pd.DataFrame): The annual or monthly natural gas pricing data.
-
-    Returns:
-        pd.DataFrame | None: Returns back the monthly data if the original data have either
-            1 or 12 data entries, otherwise None is returned.
-    """
-    match df.shape[0]:
-        case 12:
-            return df.resample("MS").bfill()  # ensure it's start of the month
-        case 1:
-            year = df.index.year[0]
-            ix = pd.date_range(f"{year}-01", f"{year}-12", freq="MS")
-            df = df.reindex(ix, method="nearest")
-            return df
-        case _:
-            pass
-
-
-def get_state_from_coords(latitude: float, longitude: float) -> str:
-    """Reverse geocodes a :py:attr:`latitude` and :py:attr:`longitude` pair to get the
-    state containing the coordinate pair.
-
-    Args:
-        latitude (float): Site latitude.
-        longitude (float): Site longitude.
-
-    Returns:
-        str: 2-letter state code (i.e., "Alabama" -> "AL").
-    """
-    try:
-        import reverse_geocoder as rg
-    except ModuleNotFoundError as e:
-        msg = (
-            "EIA natural gas feedstock coordinate input requires `reverse_geocoder` to be"
-            " installed. Directly `pip install reverse_geocoder` or use"
-            " `pip install h2integrate[gis]`."
-        )
-        raise ModuleNotFoundError(msg) from e
-
-    result = rg.search((latitude, longitude))[0]
-    return convert_state_to_code(convert_state_value(result["admin1"]))
-
-
-def convert_state_value(state: str) -> str:
-    """Convert potential two-letter state abbreviations to upper case and all else to title
-    casing to align with the ``STATE_MAP`` keys and values.
-
-    Args:
-        state (str): Either a two-letter state abbreviation or full state name.
-
-    Returns:
-        str: Upper case state abbreviation or title case state name.
-    """
-    if len(state) == 2:
-        return state.upper()
-    return state.title()
-
-
-def convert_state_to_code(state: str) -> str:
-    """Converts the :py:attr:`state` name to a two-letter abbreviation or returns the input value.
-
-    Args:
-        state (str): Full state name in title casing or two-letter state abbreviation in upper case.
-
-    Returns:
-        str: Two-letter state abbreviation.
-    """
-    return STATE_MAP.get(state, state)
-
-
-def get_eia_api_key(api_key_file: Path | None) -> str:
-    """Retrieves the EIA API key from a file, and returns the key following "EIA_API_KEY:".
-
-    Args:
-        api_key_file (Path, optional): Full file path and name of where the EIA API key is located.
-            If none is provided, then the API key is retrieved from the environment variables. Must
-            be encoded as "EIA_API_KEY: xxxxxx"
-
-    Raises:
-        ValueError: Raised either if no file is provided and an environment variable has not be
-            defined, or if a filename is provided but "EIA_API_KEY" is not found.
-
-    Returns:
-        str: The EIA API key.
-    """
-    if api_key_file is None:
-        key = os.environ.get("EIA_API_KEY")
-        if key is None:
-            msg = (
-                "No `api_key_file` provided for the EIA API, and 'EIA_API_KEY' is not defined as an"
-                " environment variable."
-            )
-            raise ValueError(msg)
-        return key
-
-    with api_key_file.open() as f:
-        for line in f.readlines():
-            if ":" in line:
-                name, val = line.strip().split(":")
-                if name == "EIA_API_KEY":
-                    return val.strip()
-    raise ValueError(f"No 'EIA_API_KEY' defined in {api_key_file=}")
 
 
 @define(kw_only=True)
